backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
pydantic==2.10.6
pydantic-settings==2.8.0
httpx==0.28.1
qdrant-client==1.13.2
supabase==2.13.0
python-dotenv==1.0.1
python-multipart==0.0.20
sentence-transformers==3.4.1
scikit-learn==1.6.1
email-validator==2.2.0
pytest==8.3.4
pytest-asyncio==0.25.3
